+++ assets/js/admin-builder-captcha.js
@@ -1,4 +1,4 @@
-/* global wpforms_builder_custom_captcha, wpforms_builder */
+/* global wpforms_builder_custom_captcha, wpforms_builder, WPFormsBuilder */
 
 /**
  * WPForms Custom Captcha admin builder function.
@@ -18,6 +18,13 @@ var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) 
 	 * @type {object}
 	 */
 	var app = {
+
+		/**
+		 * Document events on which removeEmptyQuestions() should be fired.
+		 *
+		 * @since 1.6.0
+		 */
+		removeEmptyQuestionsEvents: 'wpformsPanelSwitch wpformsFieldTabToggle wpformsFieldOptionGroupToggle',
 
 		/**
 		 * Start the engine.
@@ -48,7 +55,16 @@ var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) 
 				.on( 'click', '.wpforms-field-option-row-questions .remove', app.removeQuestion )
 
 				// Captcha sample question update.
-				.on( 'input', '.wpforms-field-option-row-questions .question', _.debounce( app.updateQuestion, 300 ) );
+				.on( 'input', '.wpforms-field-option-row-questions .question', app.updateQuestion )
+
+				// Captcha sample answer update.
+				.on( 'input', '.wpforms-field-option-row-questions .answer', app.updateAnswer )
+
+				// Remove empty questions before saving the form.
+				.on( 'wpformsBeforeSave', app.removeEmptyQuestions );
+
+			// Validate questions before panel switch and field tab toggle.
+			$( document ).on( app.removeEmptyQuestionsEvents, app.removeEmptyQuestions );
 		},
 
 		/**
@@ -131,25 +147,90 @@ var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) 
 		},
 
 		/**
+		 * Remove questions with empty question or answer values before saving the form.
+		 *
+		 * @since 1.5.0
+		 *
+		 * @param {Event} event The `wpformsBeforeSave` event object.
+		 */
+		removeEmptyQuestions: function( event ) {
+
+			const $captchaFields = $( '.wpforms-field-option-captcha' );
+
+			$captchaFields.each( function( _fieldIndex, field ) {
+
+				const id               = $( field ).data( 'field-id' );
+				const $choicesList     = $( '#wpforms-field-option-row-' + id + '-questions .choices-list' );
+				const $choices         = $choicesList.find( 'li' );
+				const $notEmptyChoices = app.getNotEmptyChoices( $choices );
+
+				// If all choices are considered empty, show and alert and stop.
+				if ( $notEmptyChoices.length === 0 ) {
+
+					app.showAlert( function() {
+
+						const currentAlert = this.$el;
+
+						// Close any other popups if active. The main goal is to close Marketing Panel popup.
+						$( '.jconfirm' ).map( function() {
+							const $jConfirm = $( this );
+							if ( ! $( currentAlert ).is( $jConfirm ) ) {
+								$jConfirm.find( '.btn-default' ).trigger( 'click' );
+							}
+						} );
+
+						$( document ).off( app.removeEmptyQuestionsEvents, app.removeEmptyQuestions );
+
+						// We're not on Fields panel, activate it.
+						if ( $( '#wpforms-panels-toggle .active' ).data( 'panel' ) !== 'fields' ) {
+							WPFormsBuilder.panelSwitch( 'fields' );
+						}
+
+						// Activate Field Options > General.
+						WPFormsBuilder.fieldTabToggle( id );
+						$( '#wpforms-field-option-' + id + ' .wpforms-field-option-group' ).removeClass( 'active' );
+						$( '#wpforms-field-option-basic-' + id ).addClass( 'active' );
+
+						$( document ).on( app.removeEmptyQuestionsEvents, app.removeEmptyQuestions );
+					} );
+
+					// Stop saving the form.
+					event.preventDefault();
+
+					return;
+				}
+
+				// We're good to go, let's remove empty choices.
+				$choices.each( function( _choiceIndex, choice ) {
+
+					const $choice = $( choice );
+
+					if ( app.isEmptyChoice( $choice ) ) {
+						$choice.remove();
+					}
+				} );
+
+				// Update preview, if needed.
+				$( '#wpforms-field-' + id ).find( '.wpforms-question' ).text( $choices.find( '.question' ).val() );
+			} );
+		},
+
+		/**
 		 * Captcha sample question update event handler.
 		 *
 		 * @since 1.3.2
 		 */
 		updateQuestion: function() {
 
-			var $this        = $( this ),
+			const $this      = $( this ),
 				$choicesList = $this.closest( '.choices-list' ),
 				$questions   = $choicesList.find( '.question' ),
 				fieldID      = $choicesList.data( 'field-id' ),
-				total        = app.getTotalNotEmptyQuestions( $questions ),
-				value        = $this.val().trim(),
-				prevValue    = $this.data( 'prev-value' ) || '';
+				value        = $this.val().trim();
 
-			if ( ! value.length && total < 1 ) {
-				app.showAlert( function() {
-					$this.val( prevValue );
-				} );
+			$this.toggleClass( 'wpforms-error', ! value.length );
 
+			if ( ! value.length ) {
 				return;
 			}
 
@@ -161,7 +242,20 @@ var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) 
 		},
 
 		/**
-		 * Show alert notifying that at least one not empty choice should remains.
+		 * Captcha sample answer update event handler.
+		 *
+		 * @since 1.6.0
+		 */
+		updateAnswer: function() {
+
+			const $this      = $( this ),
+				value        = $this.val().trim();
+
+			$this.toggleClass( 'wpforms-error', ! value.length );
+		},
+
+		/**
+		 * Show alert notifying that at least one not empty choice should remain.
 		 *
 		 * @since 1.3.2
 		 *
@@ -186,7 +280,7 @@ var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) 
 		},
 
 		/**
-		 * Show alert notifying that at least one not empty choice should remains.
+		 * Count all non-empty questions.
 		 *
 		 * @since 1.3.2
 		 *
@@ -200,6 +294,38 @@ var WPFormsCaptcha = window.WPFormsCaptcha || ( function( document, window, $ ) 
 
 				return this.value.trim().length;
 			} ).length;
+		},
+
+		/**
+		 * Filter out all choices that have either empty question or empty answer.
+		 *
+		 * @since 1.5.0
+		 *
+		 * @param {jQuery} $choices All question/answer choices.
+		 *
+		 * @returns {jQuery} Choices containing only non-empty question and answer.
+		 */
+		getNotEmptyChoices( $choices ) {
+
+			return $choices.filter( function() {
+
+				return ! app.isEmptyChoice( $( this ) );
+			} );
+		},
+
+		/**
+		 * Determine whether a question/answer choice is empty.
+		 * A choice is considered "empty" if either question or answer is empty.
+		 *
+		 * @since 1.5.0
+		 *
+		 * @param {jQuery} $choice List item containing both question and answer fields.
+		 *
+		 * @returns {boolean} Whether the choice is empty.
+		 */
+		isEmptyChoice( $choice ) {
+
+			return $choice.find( '.question' ).val().trim().length === 0 || $choice.find( '.answer' ).val().trim().length === 0;
 		},
 	};